refactor: Extract GroupPanel into a folder
