+++ src/ducks/balance/GroupPanel/GroupPanel.styl
